+++ reapy/__init__.py
@@ -24,7 +24,7 @@ def is_inside_reaper():
 
 from .tools import (
     connect, connect_to_default_machine, dist_api_is_enabled, inside_reaper,
-    reconnect
+    reconnect, is_connected
 )
 from . import reascript_api
 from .core import *