+++ reapy/tools/__init__.py
@@ -2,4 +2,5 @@
 
 import reapy
 from ._inside_reaper import inside_reaper, dist_api_is_enabled
-from .network.machines import connect, connect_to_default_machine, reconnect
+from .network.machines import (
+    connect, connect_to_default_machine, reconnect, is_connected)